Improve event timestamp resolution
Was seconds, is now milliseconds rounded to 2 (what Firefox does by default).

The specification is for 'high precision', but browsers reduce that for
fingerprinting.

The existing resolution (second), causes Browser to crash in WPT tests.

+++ src/events/event.h
@@ -32,7 +32,7 @@ struct dom_event {
 	dom_event_flow_phase phase;		/**< The event phase */
 	bool bubble;	/**< Whether this event is a bubbling event */
 	bool cancelable;	/**< Whether this event is cancelable */
-	unsigned int timestamp;
+	uint64_t timestamp;
 			/**< The timestamp this event is created */
 
 	dom_string *namespace;	/**< The namespace of this event */
